Add eip-1271 support
